feat: reorganizar configuración de RabbitMQ
