+++ src/main/java/com/facturemanagement/infraestructure/adapters/config/rabbitConfig/RabbitCommonConfig.java
@@ -1,8 +1,7 @@
-package com.facturemanagement.infraestructure.adapters.config;
+package com.facturemanagement.infraestructure.adapters.config.rabbitConfig;
 
 import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.amqp.rabbit.listener.RabbitListenerContainerFactory;
 import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer;
 import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
@@ -18,32 +17,6 @@ public class RabbitCommonConfig {
     @Bean
     Jackson2JsonMessageConverter jsonMessageConverter() {
         return new Jackson2JsonMessageConverter();
-    }
-
-    // Custom RabbitTemplate with proper error handling
-    @Bean
-    RabbitTemplate rabbitTemplate(ConnectionFactory connectionFactory) {
-        RabbitTemplate template = new RabbitTemplate(connectionFactory);
-        template.setMessageConverter(jsonMessageConverter());
-        template.setMandatory(true);
-        
-        // Return callback para mensajes no enrutables
-        template.setReturnsCallback(returned -> {
-            log.error("Message returned: {}", returned.getMessage());
-            log.error("Reply code: {}", returned.getReplyCode());
-            log.error("Reply text: {}", returned.getReplyText());
-            log.error("Exchange: {}", returned.getExchange());
-            log.error("Routing key: {}", returned.getRoutingKey());
-        });
-        
-        // Confirm callback para publishers
-        template.setConfirmCallback((correlationData, ack, cause) -> {
-            if (!ack) {
-                log.error("Message not delivered to exchange. Cause:: {}", cause);
-            }
-        });
-        
-        return template;
     }
 
     // Custom listener container factory
